pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bbssite.settings"

[tool.black]
line-length = 120
target_version = ['py39']

[tool.isort]
profile = "black"
line_length = 120

[tool.poetry]
name = "bbs"
version = "0.1.0"
description = ""
authors = ["Nikolai R Kristiansen <nikolaik@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django-extensions = "^3.1.3"
Django = "^3.2.8"
easy-thumbnails = "2.7.2"
gunicorn = "^20.1.0"
mysqlclient = "^2.0.3"
Pillow = "^9.0.0"
sentry-sdk = "1.4.3"
strawberry-graphql = "^0.93.1"
strawberry-graphql-django = "^0.2.5"
django-cors-headers = "^3.10.0"
django-rename-app = "^0.1.3"

[tool.poetry.dev-dependencies]
pytest = "*"
black = "*"
pytest-django = "*"
pre-commit = "*"
isort = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
